+++ analytics.js
@@ -2311,9 +2311,7 @@ module.exports = Provider.extend({
     authToken : null,
     // Whether or not to track an initial pageview when the page first
     // loads. You might not want this if you're using a single-page app.
-    initialPageview : true,
-    // Whether to track log calls, prefixing them with "Log: ".
-    log : false
+    initialPageview : true
   },
 
 
@@ -2346,19 +2344,6 @@ module.exports = Provider.extend({
   // If `url` is undefined, Bitdeli uses the current page URL instead.
   pageview : function (url) {
     window._bdq.push(['trackPageview', url]);
-  },
-
-  log : function (error, properties) {
-    if (!this.options.log) return;
-
-    // Bitdeli only tracks messages, so turn Error's into strings.
-    // TODO: make this cross-browser.
-    if (error instanceof Error) error = error.message;
-
-    // Prepend log, incase they're also tracking regular events.
-    error = 'Log: ' + error;
-
-    this.track(error, properties);
   }
 
 });
@@ -3204,8 +3189,12 @@ module.exports = Provider.extend({
       settings.name = traits.firstName + ' ' + traits.lastName;
     }
 
-    // Pull out a company field.
+    // Pull out a company field, with it's own optional `created` date.
     if (traits && traits.company) {
+      if (traits.company.created) {
+        traits.company.created_at = Math.floor(traits.company.created/1000);
+        delete traits.company.created;
+      }
       settings.company = traits.company;
       delete traits.company;
     }
@@ -3256,9 +3245,7 @@ module.exports = Provider.extend({
     // Whether or not to pass pageviews on to Keen IO.
     pageview : false,
     // Whether or not to track an initial pageview on `initialize`.
-    initialPageview : false,
-    // Whether to track log calls, prefixing them with "Log: ".
-    log : false
+    initialPageview : false
   },
 
   initialize : function (options, ready) {
@@ -3298,19 +3285,6 @@ module.exports = Provider.extend({
     if (url) properties = { url : url };
 
     this.track('Loaded a Page', properties);
-  },
-
-  log : function (error, properties) {
-    if (!this.options.log) return;
-
-    // Keen IO only tracks messages, so turn Error's into strings.
-    // TODO: make this cross-browser.
-    if (error instanceof Error) error = error.message;
-
-    // Prepend log, incase they're also tracking regular events.
-    error = 'Log: ' + error;
-
-    this.track(error, properties);
   }
 
 });