+++ test/providers/keen-io.js
@@ -140,31 +140,4 @@ describe('Keen IO', function () {
 
   });
 
-
-  describe('log', function () {
-
-    it('shouldnt track a log', function () {
-      var keen = analytics.providers[0]
-        , spy  = sinon.spy(keen, 'track');
-
-      analytics.log('something');
-      expect(spy.calledWith('Log: something')).to.be(false);
-      spy.restore();
-    });
-
-    it('should track a log', function () {
-      var keen = analytics.providers[0]
-        , spy  = sinon.spy(keen, 'track');
-
-      keen.options.log = true;
-
-      analytics.log('something');
-      expect(spy.calledWith('Log: something')).to.be(true);
-
-      spy.restore();
-      keen.options.log = false;
-    });
-
-  });
-
 });