+++ History.md
@@ -1,6 +1,6 @@
 0.8.10 / 2013-3-14
 ==================
-* Added conversion of company dates
+* Added fix for conversion of `company`'s `created` date
 
 0.8.9 / 2013-3-13
 =================